refactor: move mutable list
