+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/list/MutableListExtTest.kt
@@ -1,8 +1,8 @@
 package xyz.lbres.kotlinutils.list
 
-import xyz.lbres.kotlinutils.list.mutablelist.mapInPlace
-import xyz.lbres.kotlinutils.list.mutablelist.mapInPlaceIndexed
-import xyz.lbres.kotlinutils.list.mutablelist.setAllValues
+import xyz.lbres.kotlinutils.list.mapInPlace
+import xyz.lbres.kotlinutils.list.mapInPlaceIndexed
+import xyz.lbres.kotlinutils.list.setAllValues
 import xyz.lbres.kotlinutils.utils.simpleIf
 import kotlin.test.Test
 import kotlin.test.assertEquals
